Changed codegen to use scalar types for *Properties:     introduce java scalar types functions;     update writePropertiesTo and readPropertiesFrom;     preserve boxed scalars on constructor to indicate 'unset' by null;     adjusted one test to expect a scalar.

+++ test/src/com/rabbitmq/client/test/ClonePropertiesTest.java
@@ -54,7 +54,7 @@ public class ClonePropertiesTest extends TestCase {
         assertEquals(MessageProperties.MINIMAL_PERSISTENT_BASIC.getDeliveryMode(),
                      ((BasicProperties) MessageProperties.MINIMAL_PERSISTENT_BASIC.clone())
                        .getDeliveryMode());
-        assertEquals((Integer) 2,
+        assertEquals(2,
                      ((BasicProperties) MessageProperties.MINIMAL_PERSISTENT_BASIC.clone())
                        .getDeliveryMode());
     }
